Prevent compiler warning when compiling for 64 bit

+++ src/dictionary.h
@@ -40,7 +40,7 @@
 /*-------------------------------------------------------------------------*/
 typedef struct _dictionary_ {
     int             n ;     /** Number of entries in dictionary */
-    int             size ;  /** Storage size */
+    ssize_t         size ;  /** Storage size */
     char        **  val ;   /** List of string values */
     char        **  key ;   /** List of string keys */
     unsigned     *  hash ;  /** List of hash values for keys */
